+++ mobile/android/android-components/components/service/sync-logins/src/main/java/mozilla/components/service/sync/logins/AsyncLoginsStorage.kt
@@ -16,6 +16,7 @@ import mozilla.appservices.logins.MemoryLoginsStorage
 import mozilla.components.concept.sync.SyncAuthInfo
 import mozilla.components.concept.sync.SyncStatus
 import mozilla.components.concept.sync.SyncableStore
+import mozilla.appservices.sync15.SyncTelemetryPing
 
 /**
  * This type contains the set of information required to successfully
@@ -27,6 +28,11 @@ typealias SyncUnlockInfo = mozilla.appservices.logins.SyncUnlockInfo
  * Raw password data that is stored by the storage implementation.
  */
 typealias ServerPassword = mozilla.appservices.logins.ServerPassword
+
+/**
+ * The telemetry ping from a successful sync
+ */
+typealias SyncTelemetryPing = mozilla.appservices.sync15.SyncTelemetryPing
 
 /**
  * The base class of all errors emitted by logins storage.
@@ -138,7 +144,7 @@ interface AsyncLoginsStorage : AutoCloseable {
      * @rejectsWith [LoginsStorageException] if the storage is locked, and on unexpected
      *              errors (IO failure, rust panics, etc)
      */
-    fun sync(syncInfo: SyncUnlockInfo): Deferred<Unit>
+    fun sync(syncInfo: SyncUnlockInfo): Deferred<SyncTelemetryPing>
 
     /**
      * Delete all login records. These deletions will be synced to the server on the next call to sync.
@@ -302,7 +308,7 @@ open class AsyncLoginsStorageAdapter<T : LoginsStorage>(private val wrapped: T) 
         return wrapped.isLocked()
     }
 
-    override fun sync(syncInfo: SyncUnlockInfo): Deferred<Unit> {
+    override fun sync(syncInfo: SyncUnlockInfo): Deferred<SyncTelemetryPing> {
         return scope.async {
             wrapped.sync(syncInfo)
         }